+++ src/components/Ads/index.ts
@@ -1,0 +1,2 @@
+export { AdContainer } from './AdContainer';
+export type { OwAd } from "@overwolf/types/owads";